% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_nights_weekends_schedules.R
\name{add_nights_weekends_schedules}
\alias{add_nights_weekends_schedules}
\title{Join Nights and Weekends Schedules to Teams Data}
\usage{
add_nights_weekends_schedules(
  .data,
  path = path_create("V:/Administration/Schedules/Night&Weekend staff", ext = "xlsx"),
  sheet = "schedules"
)
}
\arguments{
\item{.data}{The output of \code{as_join_schedules()}}

\item{path}{The location of the nights and weekends scheduling Excel workbook}

\item{sheet}{The sheet from which to read the schedule; either the sheet name
or the sheet number}
}
\value{
A \code{tibble} with a \code{schedule_custom} column added
}
\description{
\code{add_nights_weekends_schedules()} loads individual schedules for
investigators on night or weekend shifts and joins them with the output of
\code{join_schedules()}.
}
\seealso{
Other Case Assignment: 
\code{\link{calc_schedules}()},
\code{\link{join_schedules}()},
\code{\link{load_nights_weekends}()},
\code{\link{load_teams}()},
\code{\link{parse_nights_weekends}()},
\code{\link{schedule_predefined}()},
\code{\link{schedule}()},
\code{\link{std_names}()}
}
\concept{Case Assignment}
