% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_acns.R
\name{archive_translated}
\alias{archive_translated}
\title{Archive Translated ACNS Data}
\usage{
archive_translated(
  data,
  dir = "V:/EPI DATA ANALYTICS TEAM/Case Assignment/data/archive/new_tests",
  force = FALSE
)
}
\arguments{
\item{data}{Data from \code{prep_acns(assign = TRUE)}}

\item{dir}{Archive directory}

\item{force}{Should an existing file be overwritten?}
}
\value{
The path to the archive file
}
\description{
\code{archive_translated()} archives cases translated from the output of
\code{\link[covidsms:prep_acns]{prep_acns(assign = TRUE)}}. It saves the
translated data without deduplicating.
}
