% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_schedules.R
\name{asg_calc_schedules}
\alias{asg_calc_schedules}
\title{Calculate Whether Investigators Are Working on a Given Date}
\usage{
asg_calc_schedules(.data, date = Sys.Date())
}
\arguments{
\item{.data}{The output of \code{asg_join_nights_weekends()}}

\item{date}{The date to use when checking whether investigators are scheduled}
}
\value{
A \code{tibble} with columns \code{team}, \code{member}, and \code{scheduled}
}
\description{
\code{asg_calc_schedules()} takes the output of
\code{asg_add_nights_weekends_schedules()} and calcuates whether investigators are
scheduled to work on \code{date}. Either a predefined schedule or a custom
schedule must be supplied for every investigator.
}
\seealso{
Other Case Assignment: 
\code{\link{asg_add_nights_weekends_schedules}()},
\code{\link{asg_join_schedules}()},
\code{\link{asg_load_nights_weekends}()},
\code{\link{asg_load_teams}()},
\code{\link{asg_parse_nights_weekends}()},
\code{\link{asg_schedule_predefined}()},
\code{\link{asg_schedule}()},
\code{\link{asg_std_names}()}
}
\concept{Case Assignment}
