% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_schedules.R
\name{asg_join_schedules}
\alias{asg_join_schedules}
\title{Join the \code{team_schedules} Dataset to Teams}
\usage{
asg_join_schedules(teams, schedules = team_schedules)
}
\arguments{
\item{teams}{A data frame containing teams in columns and members in rows;
designed to take the output of
\code{\link[coviData:asg_parse_teams]{asg_parse_teams()}}}

\item{schedules}{Schedules for each team, given by the \code{team_schedules}
dataset}
}
\value{
A \code{tibble} with columns for \code{team}, \code{member}, \code{schedule}, and
\code{anchor}
}
\description{
\code{asg_join_schedules()} pivots the \code{teams} data to long format and joins the
schedules specified in \code{\link{team_schedules}} to the result.
}
\seealso{
Other Case Assignment: 
\code{\link{asg_add_nights_weekends_schedules}()},
\code{\link{asg_calc_schedules}()},
\code{\link{asg_load_nights_weekends}()},
\code{\link{asg_load_teams}()},
\code{\link{asg_parse_names}()},
\code{\link{asg_parse_nights_weekends}()},
\code{\link{asg_schedule_predefined}()},
\code{\link{asg_schedule}()}
}
\concept{Case Assignment}
