% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_nights_weekends.R
\name{asg_parse_nights_weekends}
\alias{asg_parse_nights_weekends}
\title{Parse and Re-Shape Nights and Weekends Scheduling Data}
\usage{
asg_parse_nights_weekends(.data)
}
\arguments{
\item{.data}{Data read by
\code{\link[covidata:asg_load_nights_weekends]{asg_load_nights_weekends()}}}
}
\value{
A \code{tibble} with one row per investigator and columns named \code{member}
and \code{schedule}; the latter is a list-column containing named logical
vectors of the weekly schedules for each investigator
}
\description{
\code{asg_parse_nights_weekends()} parses night/weekend schedules into a standard
format for case assignment.
}
\seealso{
Other Case Assignment: 
\code{\link{asg_add_nights_weekends_schedules}()},
\code{\link{asg_calc_schedules}()},
\code{\link{asg_join_schedules}()},
\code{\link{asg_load_nights_weekends}()},
\code{\link{asg_load_teams}()},
\code{\link{asg_parse_names}()},
\code{\link{asg_schedule_predefined}()},
\code{\link{asg_schedule}()}
}
\concept{Case Assignment}
