% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_teams.R
\name{asg_parse_teams}
\alias{asg_parse_teams}
\title{Parse the Team Schedule Workbook to Standardized Format}
\usage{
asg_parse_teams(.data)
}
\arguments{
\item{.data}{The teams data read by \code{asg_load_teams()}}
}
\value{
A \code{tibble} containing one column for each team (named with that
team's letter designation) and one row per team member. Team member names
are standardized with
\code{\link[covidassign:asg_std_names]{asg_parse_names()}}.
}
\description{
\code{asg_parse_teams()} parses the data returned from \code{asg_load_teams()} into
a standard format, with one column per team and one row per individual.
}
