% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedule.R, R/schedule_by_cycle.R,
%   R/schedule_by_day.R
\name{asg_schedule}
\alias{asg_schedule}
\alias{asg_schedule_by_cycle}
\alias{asg_schedule_by_day}
\title{Define Rotating or Weekly Work Schedules}
\usage{
asg_schedule(
  cycle = c(Sun = FALSE, Mon = TRUE, Tue = TRUE, Wed = TRUE, Thu = TRUE, Fri = TRUE,
    Sat = FALSE),
  start = Sys.Date(),
  end = Sys.Date() + 29L,
  anchor = start
)

asg_schedule_by_cycle(
  cycle = c(FALSE, TRUE, TRUE, TRUE, TRUE, TRUE, FALSE),
  start = Sys.Date(),
  end = Sys.Date() + 29L,
  anchor = start
)

asg_schedule_by_day(
  cycle = c(Sun = FALSE, Mon = TRUE, Tue = TRUE, Wed = TRUE, Thu = TRUE, Fri = TRUE,
    Sat = FALSE),
  start = Sys.Date(),
  end = Sys.Date() + 29L
)
}
\arguments{
\item{cycle}{A logical vector defining one scheduling cycle. If named, names
are passed to \code{\link[covidassign:parse_weekday]{parse_weekday()}} for
standardization of weekday names.}

\item{start}{The start date of the returned schedule; either a string in
"YYYY-MM-DD" format or a \code{Date} object}

\item{end}{The end date of the returned schedule; either a string in
"YYYY-MM-DD" format or a \code{Date} object}

\item{anchor}{The date from which to start ("anchor") schedule calculations.
This can be any valid date; no particular relationship to \code{start} or \code{end}
is needed. It must be either a string in "YYYY-MM-DD" format or a
\code{Date} object.}
}
\value{
A \code{tibble} with columns \code{date} (a \code{Date} column containing dates
between \code{start} and \code{end}, inclusive), \code{weekday} (a \code{character} column
containing full weekday names), and \code{scheduled} (a \code{logical} column
defining whether a day is schedule ("on") or not ("off"))
}
\description{
\code{asg_schedule()} defines work schedules based off of a given \code{cycle}. If this
cycle is named, it is assumed that the names are days of the week, and that
the schedule repeats weekly. If it is unnamed, the \code{cycle} begins at the
\code{anchor} date. Defining rotating weekly schedules
(i.e. T/R one week, M/W/F the next) is not currently supported using weekday
names; these schedules may be defined as a rotating schedule anchored to the
beginning of a week.
}
\details{
\code{asg_schedule_by_cycle()} and \code{asg_schedule_by_day()} are the workhorses
underlying \code{as_schedule()}. They handle the general cyclic and weekly use
cases described above.
}
\seealso{
Other Case Assignment: 
\code{\link{asg_add_nights_weekends_schedules}()},
\code{\link{asg_calc_schedules}()},
\code{\link{asg_join_schedules}()},
\code{\link{asg_load_nights_weekends}()},
\code{\link{asg_load_teams}()},
\code{\link{asg_parse_nights_weekends}()},
\code{\link{asg_schedule_predefined}()},
\code{\link{asg_std_names}()}
}
\concept{Case Assignment}
