% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/std_names.R
\name{asg_std_names}
\alias{asg_std_names}
\title{Parse Investigator Names and Remove (Most) Non-Alphabetical Characters}
\usage{
asg_std_names(string)
}
\arguments{
\item{string}{A character vector containing investigator names}
}
\value{
The input \code{string} with cleaned names
}
\description{
\code{asg_std_names()} parses investigator names given in the teams
worksheet into a standard format for matching with REDcap. It:
\enumerate{
\item Replaces square brackets and curly braces with parentheses
\item Removes parenthetic expressions
\item Removes ordinal and numeric expressions
\item Converts \code{NA}s to text (for the next function)
\item Passes the intermediate to
\code{\link[janitor:make_clean_names]{
      make_clean_names(
        case = "title",
        use_make_names = FALSE,
        transliterations = c("Any-Latin", "Latin-ASCII"),
        sep_out = " "
      )
    }}
\item Removes leading and trailing dashes
\item Removes extraneous whitespace
\item Inverts converted \code{NA}s (back to \code{NA_character_})
}
}
\seealso{
Other Case Assignment: 
\code{\link{asg_add_nights_weekends_schedules}()},
\code{\link{asg_calc_schedules}()},
\code{\link{asg_join_schedules}()},
\code{\link{asg_load_nights_weekends}()},
\code{\link{asg_load_teams}()},
\code{\link{asg_parse_nights_weekends}()},
\code{\link{asg_schedule_predefined}()},
\code{\link{asg_schedule}()}
}
\concept{Case Assignment}
