% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_cases.R
\name{assign_cases}
\alias{assign_cases}
\title{Assign Cases to Scheduled Investigators}
\usage{
assign_cases(date = Sys.Date(), api_token = Sys.getenv("redcap_CA_token"))
}
\arguments{
\item{date}{The date to use when checking whether investigators are scheduled}

\item{api_token}{The API token for the Case Assignment REDcap project. The
default pulls this from the environment variable \code{redcap_CA_token}.}
}
\value{
A \code{tibble} containing the results of
\code{
  \link[covidassign:download_redcap_cases]{download_redcap_cases()}
  }, with
\code{assign_date} and \code{investigator} filled
}
\description{
\code{assign_cases()} divides unassigned cases among REDcap investigators
scheduled to work on \code{date}. It distributes cases randomly and evenly across
investigators until either all are assigned or there are too few cases for
another round of even distribution. Any remaining cases are assigned randomly
with at most one additional case given to an investigator.
}
