% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_acns.R
\name{assign_general}
\alias{assign_general}
\title{Assign General Cases to Investigators}
\usage{
assign_general(
  .data,
  date = attr(.data, "date"),
  api_token = Sys.getenv("redcap_NCA_token")
)
}
\arguments{
\item{.data}{Data frame of cases to assign, passed from
\code{\link[covidassign:translate_acns]{translate_acns()}}}

\item{date}{The date to use when checking whether investigators are scheduled}

\item{api_token}{The API token for the Case Assignment REDcap project. The
default pulls this from the environment variable \code{redcap_CA_token}.}
}
\value{
The input \code{.data} with general cases assigned via \code{investigator},
\code{team}, and \code{assign_date} columns
}
\description{
Assign General Cases to Investigators
}
