% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_acns.R
\name{assign_long_term}
\alias{assign_long_term}
\title{Assign Long Term Care Facility Cases to Investigators}
\usage{
assign_long_term(
  .data,
  assign = "Hawa Abdalla",
  api_token = Sys.getenv("redcap_NCA_token")
)
}
\arguments{
\item{.data}{Data frame of cases to assign, passed from
\code{\link[covidassign:translate_acns]{translate_acns()}}}

\item{assign}{Character vector of names of investigators to
assign long term care facility cases}

\item{api_token}{The API token for the Case Assignment REDcap project. The
default pulls this from the environment variable \code{redcap_CA_token}.}
}
\value{
The input \code{.data} with long term care facility cases assigned via
\code{investigator}, \code{team}, and \code{assign_date} columns
}
\description{
Assign Long Term Care Facility Cases to Investigators
}
