% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_acns.R
\name{assign_school_age}
\alias{assign_school_age}
\title{Assign School Age Cases to Investigators}
\usage{
assign_school_age(
  .data,
  assign = c("Reed Wagner", "Camry Gipson", "Logan Sell", "Shanna Layrock"),
  api_token = Sys.getenv("redcap_NCA_token")
)
}
\arguments{
\item{.data}{Data frame of cases to assign, passed from
\code{\link[covidassign:translate_acns]{translate_acns()}}}

\item{api_token}{The API token for the Case Assignment REDcap project. The
default pulls this from the environment variable \code{redcap_CA_token}.}
}
\value{
The input \code{.data} with school age cases assigned via \code{investigator},
\code{team}, and \code{assign_date} columns
}
\description{
Assign School Age Cases to Investigators
}
