% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_schedules.R
\name{calc_schedules}
\alias{calc_schedules}
\title{Calculate Whether Investigators Are Working on a Given Date}
\usage{
calc_schedules(.data, date = Sys.Date())
}
\arguments{
\item{.data}{The output of \code{join_nights_weekends()}}

\item{date}{The date to use when checking whether investigators are scheduled}
}
\value{
A \code{tibble} with columns \code{team}, \code{member}, and \code{scheduled}
}
\description{
\code{calc_schedules()} takes the output of
\code{add_nights_weekends_schedules()} and calcuates whether investigators are
scheduled to work on \code{date}. Either a predefined schedule or a custom
schedule must be supplied for every investigator.
}
\seealso{
Other Case Assignment: 
\code{\link{add_nights_weekends_schedules}()},
\code{\link{join_schedules}()},
\code{\link{load_nights_weekends}()},
\code{\link{load_teams}()},
\code{\link{parse_nights_weekends}()},
\code{\link{schedule_predefined}()},
\code{\link{schedule}()},
\code{\link{std_names}()}
}
\concept{Case Assignment}
