% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_redcap_investigators.R
\name{download_redcap_investigators}
\alias{download_redcap_investigators}
\title{Download Investigators Listed in the Case Assignment REDcap Project}
\usage{
download_redcap_investigators(api_token = Sys.getenv("redcap_CA_token"))
}
\arguments{
\item{api_token}{The API token for the Case Assignment REDcap project. The
default pulls this from the environment variable \code{redcap_CA_token}.}
}
\value{
A \code{tibble} with one row for each investigator and columns for \code{id}
and \code{investigator} (name)
}
\description{
\code{download_redcap_investigators()} downloads all investigators available
for assignment in the \strong{Case Assignment} REDcap project, as well as the
\code{id} assigned to each investigator in REDcap.
}
