% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_redcap_cases.R
\name{download_redcap_records}
\alias{download_redcap_records}
\alias{download_nca_records}
\alias{download_nit_records}
\title{Download Cases from Case Assignment REDcap Project}
\usage{
download_nca_records(api_token = Sys.getenv("redcap_NCA_token"), fields = NULL)

download_nit_records(api_token = Sys.getenv("redcap_NIT_token"), fields = NULL)
}
\arguments{
\item{api_token}{The API token for the Case Assignment REDcap project}

\item{fields}{\code{character}. The fields to download from the project.
By default, all fields are downloaded.}
}
\value{
A \code{tibble} with all variables in \code{character} format
}
\description{
\code{download_nit_records()} and \code{download_nca_records()} download records from
the \strong{Interview Tool} and \strong{Case Assignment} REDcap projects.
}
\section{TODO}{

\enumerate{
\item Check \code{six} column in NCA to see if exclusion is necessary
\item Create unit tests
}
}

