% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_addresses.R
\name{expand_city_dict}
\alias{expand_city_dict}
\title{Expand a City Dictionary with Approximate String Matching}
\usage{
expand_city_dict(.data, city_dict, dist_max = 2)
}
\arguments{
\item{.data}{The output of \code{prep_city()}}

\item{city_dict}{A city dictionary created by \code{pm_append()} or
\code{pm_dictionary()}}

\item{dist_max}{Maximum OSA distance to consider a match when expanding city
names}
}
\value{
A \code{tibble} with columns \code{city.input} holding variants of city names
found in the data and \code{city.output} holding standardized city names given
by \code{city_dict}
}
\description{
\code{expand_city_dict()} is an internal function used by \code{parse_addresses()}. It
takes an existing dictionary of cities and expands it to match city names
within a specified \code{\link[stringdist:stringdist-metrics]{OSA distance}}
of the \code{city.output} values. This is useful for handling typos.
}
\keyword{internal}
