% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_investigators.R
\name{get_investigators}
\alias{get_investigators}
\title{Join Investigators Listed in Investigation Schedule & Case Assignment Project}
\usage{
get_investigators(
  date = Sys.Date(),
  type = c("all", "general", "school"),
  scheduled_only = TRUE,
  api_token = Sys.getenv("redcap_NCA_token"),
  quiet = TRUE
)
}
\arguments{
\item{date}{The date to use when checking whether investigators are scheduled}

\item{type}{The type or group of investigators to return}

\item{scheduled_only}{Should only scheduled investigators be returned? If
\code{FALSE}, all investigators will be returned. Useful for debugging.}

\item{api_token}{The API token for the Case Assignment REDcap project. The
default pulls this from the environment variable \code{redcap_CA_token}.}

\item{quiet}{Should update messages be suppressed?}
}
\value{
A \code{tibble} with one row per investigator and an \code{investigator}
column. If \code{type = "inner"}, an \code{id} column is also included.
}
\description{
\code{get_investigators()} joins the investigators scheduled to work on
\code{date} with the list of investigators from the \strong{Case Assignment}
REDcap project. By default, it performs an inner join, since only
investigators in both lists are available for case assignment.
}
\details{
While the default behavior is an inner join, \code{get_investigators()} can
also perform anti-joins with either the scheduled or REDcap investigator list
as the primary table. This is useful for debugging and checking that all
scheduled investigators are able to have cases assigned.
}
