% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_schedules.R
\name{join_schedules}
\alias{join_schedules}
\title{Join the \code{team_schedules} Dataset to Teams}
\usage{
join_schedules(teams, schedules = team_schedules)
}
\arguments{
\item{teams}{A data frame containing teams in columns and members in rows;
designed to take the output of
\code{\link[covidassign:parse_teams]{parse_teams()}}}

\item{schedules}{Schedules for each team, given by the \code{team_schedules}
dataset}
}
\value{
A \code{tibble} with columns for \code{team}, \code{member}, \code{schedule}, and
\code{anchor}
}
\description{
\code{join_schedules()} pivots the \code{teams} data to long format and joins the
schedules specified in \code{\link{team_schedules}} to the result.
}
\seealso{
Other Case Assignment: 
\code{\link{add_nights_weekends_schedules}()},
\code{\link{calc_schedules}()},
\code{\link{load_nights_weekends}()},
\code{\link{load_teams}()},
\code{\link{parse_nights_weekends}()},
\code{\link{schedule_predefined}()},
\code{\link{schedule}()},
\code{\link{std_names}()}
}
\concept{Case Assignment}
