% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_nbs_cases.R
\name{load_nbs_positive}
\alias{load_nbs_positive}
\title{Load Positive Tests From NBS Snapshot File}
\usage{
load_nbs_positive(date = NULL)
}
\arguments{
\item{date}{Optional. Character or \code{Date}. The date to look for when
selecting a PCR file to read.}
}
\description{
\code{load_nbs_positive()} loads positive tests from the latest PCR snapshot file,
or the one specified by \code{date}, if provided.
}
