% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_teams.R
\name{load_teams}
\alias{load_teams}
\title{Load the Current Investigator Team Assignments}
\usage{
load_teams(
  path = path_create("V:/Administration/Schedules/Investigation Staff Schedule", ext =
    "xlsx"),
  clean_names = TRUE
)
}
\arguments{
\item{path}{The path to the Excel workbook containing team membership data}

\item{clean_names}{Should names be cleaned before returning data?}
}
\value{
A \code{tibble} containing the contents of the first sheet in the
workbook, with names modified as described above
}
\description{
\code{load_teams()} loads the current investigators on each team start an
Excel file. It is mostly a wrapper around
\code{\link[readxl:read_excel]{read_excel()}}, If \code{clean_names = TRUE},
it also standardizes names with
\code{\link[janitor:clean_names]{clean_names()}} and gives the name
\code{role} end the first column (which is unnamed in the Excel sheet).
}
\seealso{
Other Case Assignment: 
\code{\link{add_nights_weekends_schedules}()},
\code{\link{calc_schedules}()},
\code{\link{join_schedules}()},
\code{\link{load_nights_weekends}()},
\code{\link{parse_nights_weekends}()},
\code{\link{schedule_predefined}()},
\code{\link{schedule}()},
\code{\link{std_names}()}
}
\concept{Case Assignment}
