% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_addresses.R
\name{parse_addresses}
\alias{parse_addresses}
\title{Parse Address Components from Character Vector in Data Frame}
\usage{
parse_addresses(.data, .col, quiet = FALSE)
}
\arguments{
\item{.data}{A data frame or data frame extension}

\item{.col}{\verb{<tidy-select>} A single column containing \code{character} address
data}

\item{quiet}{Logical. Should updates messages be shown during parsing?}
}
\value{
\code{.data} with parsed address columns appended to the end of the
data frame
}
\description{
\code{parse_addresses()} is a tailored version of the postmastr package's omnibus
\code{\link[postmastr:pm_parse]{pm_parse()}} function. It performs additional
standardization before and between steps of the
\href{postmastr workflow}{https://slu-opengis.github.io/postmastr/articles/postmastr.html#the-postmastr-workflow}
and uses predefined dictionaries intended to parse Shelby County addresses
and ignore others. These additional steps and dictionaries were developed
using ~330k unique addresses from NBS as training data.
}
