% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_nights_weekends.R
\name{parse_nights_weekends}
\alias{parse_nights_weekends}
\title{Parse and Re-Shape Nights and Weekends Scheduling Data}
\usage{
parse_nights_weekends(.data)
}
\arguments{
\item{.data}{Data read by
\code{
  \link[covidassign:load_nights_weekends]{load_nights_weekends()}
  }}
}
\value{
A \code{tibble} with one row per investigator and columns named \code{member}
and \code{schedule}; the latter is a list-column containing named logical
vectors of the weekly schedules for each investigator
}
\description{
\code{parse_nights_weekends()} parses night/weekend schedules into a standard
format for case assignment.
}
\seealso{
Other Case Assignment: 
\code{\link{add_nights_weekends_schedules}()},
\code{\link{calc_schedules}()},
\code{\link{join_schedules}()},
\code{\link{load_nights_weekends}()},
\code{\link{load_teams}()},
\code{\link{schedule_predefined}()},
\code{\link{schedule}()},
\code{\link{std_names}()}
}
\concept{Case Assignment}
