% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_teams.R
\name{parse_teams}
\alias{parse_teams}
\title{Parse the Team Schedule Workbook to Standardized Format}
\usage{
parse_teams(.data)
}
\arguments{
\item{.data}{The teams data read by \code{load_teams()}}
}
\value{
A \code{tibble} containing one column for each team (named with that
team's letter designation) and one row per team member. Team member names
are standardized with
\code{\link[covidassign:std_names]{parse_names()}}.
}
\description{
\code{parse_teams()} parses the data returned from \code{load_teams()} into
a standard format, with one column per team and one row per individual.
}
