% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_by_pattern.R
\name{path_teams}
\alias{path_teams}
\alias{path_nights_weekends}
\alias{path_by_pattern}
\title{Get Path(s) to the Most Recent Scheduling Files}
\usage{
path_teams(
  dir = "V:/Administration/Schedules/",
  pattern = "(?i).*/[^/~]*Investigations?[ ]*Staff[ ]*Schedule.*xlsx?$"
)

path_nights_weekends(
  dir = "V:/Administration/Schedules/",
  pattern = "(?i).*/[^/~]*Nights?.*Weekends?[ ]*Staff.*xlsx?$"
)

path_by_pattern(dir, pattern)
}
\arguments{
\item{dir}{The path to the directory of interest}

\item{pattern}{The regular expression to use when filtering paths}
}
\value{
An \code{fs_path} vector
}
\description{
\code{path_teams()} finds the path to the most recent investigation teams
schedule.

\code{path_nights_weekends()} finds the path to the most recent nights/weekends
schedule.
}
\details{
These functions returns the most recently created file; if multiple files are
create within 1 second of each other, this will return a vector of paths to
each of those files.

These functions are powered (and generalized) by \code{path_by_pattern()}, which
finds the most recent file in a given directory that matches a regular
expression.
}
