% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_acns_redcap.R
\name{prep_acns_redcap}
\alias{prep_acns_redcap}
\title{Prepare Assigned Cases for REDcap Upload}
\usage{
prep_acns_redcap(.data = assign_acns(), date = attr(.data, "date"))
}
\arguments{
\item{.data}{Assigned case data}

\item{date}{The date to assume for "today"; default is given by the \code{date}
attribute of the input \code{date_tbl}}
}
\value{
Data ready for upload via \code{upload_assignments()}
}
\description{
\code{prep_acns_redcap()} performs final data cleaning to prepare assigned cases
for upload to REDcap.
}
