% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_addresses.R
\name{prep_city}
\alias{prep_city}
\title{Prepare Data for Parsing by \code{pm_city_parse()}}
\usage{
prep_city(.data)
}
\arguments{
\item{.data}{The output of \code{pm_parse_state()}}
}
\value{
\code{.data} with the \code{pm.address} variable prepared for parsing by
\code{pm_parse_city()}
}
\description{
\code{prep_city()} is an internal function used by \code{parse_addresses()}. It
re-parses postal codes and states from the data to clear any double entry
due to multiple fields being concatenated, and it "squishes" extraneous
whitespace.
}
\keyword{internal}
