% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_addresses.R
\name{prep_house}
\alias{prep_house}
\title{Prepare Data for Parsing by \code{pm_house_parse()}}
\usage{
prep_house(.data, city_dict)
}
\arguments{
\item{.data}{The output of \code{pm_parse_city()}}
}
\value{
\code{.data} with the \code{pm.address} variable prepared for parsing by
\code{pm_parse_house()}
}
\description{
\code{prep_house()} is an internal function used by \code{parse_addresses()}. It first
re-parses postal codes, states, and cities to clear any double entry due to
multiple fields being concatenated. It then cleans references to
apartments/units/rooms/PO Boxes (and associated numeric expressions) and
"squishes" extraneous whitespace
}
\keyword{internal}
