% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_addresses.R
\name{prep_state}
\alias{prep_state}
\title{Prepare Data for Parsing by \code{pm_state_parse()}}
\usage{
prep_state(.data)
}
\arguments{
\item{.data}{The output of \code{pm_prep()}}
}
\value{
\code{.data} with the \code{pm.address} variable prepared for parsing by
\code{pm_parse_state()}
}
\description{
\code{prep_state()} is an internal function used by \code{parse_addresses()}. It
removes trailing numeric expressions that are not the Tennessee FIPS code
(\code{"47"}) and "squishes" extraneous whitespace.
}
\keyword{internal}
