% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_redcap_cases.R
\name{redcap_fields}
\alias{redcap_fields}
\title{Select REDcap Fields to Download}
\usage{
redcap_fields(fields)
}
\arguments{
\item{fields}{\code{character}. REDcap fields to select.}
}
\value{
A named \code{list}
}
\description{
\code{redcap_fields()} converts a \code{character} vector to a named \code{list} that can
be appended to the REDcap API parameters
}
