% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sched_add_nights_weekends.R
\name{sched_add_nights_weekends}
\alias{sched_add_nights_weekends}
\title{Join Nights and Weekends Schedules to Teams Data}
\usage{
sched_add_nights_weekends(
  .data,
  path = path_nights_weekends(),
  sheet = "schedules"
)
}
\arguments{
\item{.data}{The output of \code{sched_join_schedules()}}

\item{path}{The location of the nights and weekends scheduling Excel workbook}

\item{sheet}{The sheet from which to read the schedule; either the sheet name
or the sheet number}
}
\value{
A \code{tibble} with a \code{schedule_custom} column added
}
\description{
\code{sched_add_nights_weekends()} loads individual schedules for
investigators on night or weekend shifts and joins them with the output of
\code{sched_join_schedules()}.
}
\seealso{
Other Case Assignment: 
\code{\link{sched_calc}()},
\code{\link{sched_join_schedules}()},
\code{\link{sched_load_nights_weekends}()},
\code{\link{sched_load_teams}()},
\code{\link{sched_parse_nights_weekends}()},
\code{\link{sched_std_names}()},
\code{\link{schedule_predefined}()},
\code{\link{schedule}()}
}
\concept{Case Assignment}
