% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sched_calc.R
\name{sched_calc}
\alias{sched_calc}
\title{Calculate Whether Investigators Are Working on a Given Date}
\usage{
sched_calc(.data, date = Sys.Date())
}
\arguments{
\item{.data}{The output of \code{join_nights_weekends()}}

\item{date}{The date to use when checking whether investigators are scheduled}
}
\value{
A \code{tibble} with columns \code{team}, \code{member}, and \code{scheduled}
}
\description{
\code{sched_calc()} takes the output of
\code{sched_add_nights_weekends()} and calcuates whether investigators are
scheduled to work on \code{date}. Either a predefined schedule or a custom
schedule must be supplied for every investigator.
}
\seealso{
Other Case Assignment: 
\code{\link{sched_add_nights_weekends}()},
\code{\link{sched_join_schedules}()},
\code{\link{sched_load_nights_weekends}()},
\code{\link{sched_load_teams}()},
\code{\link{sched_parse_nights_weekends}()},
\code{\link{sched_std_names}()},
\code{\link{schedule_predefined}()},
\code{\link{schedule}()}
}
\concept{Case Assignment}
