% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sched_investigators.R
\name{sched_investigators}
\alias{sched_investigators}
\title{Get Investigators Scheduled to Work on a Given Date}
\usage{
sched_investigators(
  date = lubridate::today(),
  team_sched = FALSE,
  scheduled_only = TRUE
)
}
\arguments{
\item{date}{The date to use when checking whether investigators are scheduled}

\item{team_sched}{Should the team schedules be used as "ground truth" for
team commposition and a backup for scheduling?}

\item{scheduled_only}{Should only scheduled investigators be returned? If
\code{FALSE}, all investigators will be returned. Useful for debugging.}
}
\value{
A \code{tibble} containing \code{team}, \code{investigator}, and \code{scheduled} columns
}
\description{
\code{sched_investigators()} parses scheduling files and returns
investigators scheduled to work on \code{date}.
}
