% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sched_load_nights_weekends.R
\name{sched_load_nights_weekends}
\alias{sched_load_nights_weekends}
\title{Load Night & Weekend Scheduling Data}
\usage{
sched_load_nights_weekends(
  path = path_create("V:/Administration/Schedules/Night&Weekend staff", ext = "xlsx"),
  sheet = "schedules",
  clean_names = TRUE
)
}
\arguments{
\item{path}{The location of the nights and weekends scheduling Excel workbook}

\item{sheet}{The sheet from which to read the schedule; either the sheet name
or the sheet number}

\item{clean_names}{Should names be cleaned?}
}
\value{
The data in the read sheet as a \code{tibble}
}
\description{
\code{sched_load_nights_weekends()} loads schedules for investigators working
nights and weekends. These team members do not follow a team-wide schedule
and must be dealt with individually.
}
\seealso{
Other Case Assignment: 
\code{\link{sched_add_nights_weekends}()},
\code{\link{sched_calc}()},
\code{\link{sched_join_schedules}()},
\code{\link{sched_load_teams}()},
\code{\link{sched_parse_nights_weekends}()},
\code{\link{sched_std_names}()},
\code{\link{schedule_predefined}()},
\code{\link{schedule}()}
}
\concept{Case Assignment}
