% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sched_load_teams.R
\name{sched_load_teams}
\alias{sched_load_teams}
\title{Load the Current Investigator Team Assignments}
\usage{
sched_load_teams(path = path_teams(), clean_names = TRUE)
}
\arguments{
\item{path}{The path to the Excel workbook containing team membership data}

\item{clean_names}{Should names be cleaned before returning data?}
}
\value{
A \code{tibble} containing the contents of the first sheet in the
workbook, with names modified as described above
}
\description{
\code{sched_load_teams()} loads the current investigators on each team start an
Excel file. It is mostly a wrapper around
\code{\link[readxl:read_excel]{read_excel()}}, If \code{clean_names = TRUE},
it also standardizes names with
\code{\link[janitor:clean_names]{clean_names()}} and gives the name
\code{role} end the first column (which is unnamed in the Excel sheet).
}
\seealso{
Other Case Assignment: 
\code{\link{sched_add_nights_weekends}()},
\code{\link{sched_calc}()},
\code{\link{sched_join_schedules}()},
\code{\link{sched_load_nights_weekends}()},
\code{\link{sched_parse_nights_weekends}()},
\code{\link{sched_std_names}()},
\code{\link{schedule_predefined}()},
\code{\link{schedule}()}
}
\concept{Case Assignment}
