% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sched_load_tomorrow.R
\name{sched_load_tomorrow}
\alias{sched_load_tomorrow}
\title{Load Schedules for Tomorrow Provided by Team Leads}
\usage{
sched_load_tomorrow(path)
}
\arguments{
\item{path}{Character. Path to the excel file containing schedules.}
}
\value{
A \code{tibble} with columns \code{investigator} (\code{character}) and
\code{scheduled} (\code{logical})
}
\description{
\code{sched_load_tomorrow()} loads schedules for the following day filled out
by team leads.
}
