% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sched_load_tomorrow.R
\name{sched_load_tomorrow_all}
\alias{sched_load_tomorrow_all}
\title{Load All Schedules for Tomorrow Provided by Team Leads}
\usage{
sched_load_tomorrow_all(
  dir = "V:/Investigation Team Leads/Daily Assignment Schedule/"
)
}
\arguments{
\item{dir}{Character. Path to directory containing excel file schedules.}
}
\value{
A \code{tibble} with columns \code{investigator} (\code{character}) and
\code{scheduled} (\code{logical})
}
\description{
\code{sched_load_tomorrow_all()} loads all schedules provided by team leads.
}
