% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sched_parse_nights_weekends.R
\name{sched_parse_nights_weekends}
\alias{sched_parse_nights_weekends}
\title{Parse and Re-Shape Nights and Weekends Scheduling Data}
\usage{
sched_parse_nights_weekends(.data)
}
\arguments{
\item{.data}{Data read by
\code{
  \link[covidassign:sched_load_nights_weekends]{sched_load_nights_weekends()}
  }}
}
\value{
A \code{tibble} with one row per investigator and columns named \code{member}
and \code{schedule}; the latter is a list-column containing named logical
vectors of the weekly schedules for each investigator
}
\description{
\code{sched_parse_nights_weekends()} parses night/weekend schedules into a standard
format for case assignment.
}
\seealso{
Other Case Assignment: 
\code{\link{sched_add_nights_weekends}()},
\code{\link{sched_calc}()},
\code{\link{sched_join_schedules}()},
\code{\link{sched_load_nights_weekends}()},
\code{\link{sched_load_teams}()},
\code{\link{sched_std_names}()},
\code{\link{schedule_predefined}()},
\code{\link{schedule}()}
}
\concept{Case Assignment}
