% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sched_parse_teams.R
\name{sched_parse_teams}
\alias{sched_parse_teams}
\title{Parse the Team Schedule Workbook to Standardized Format}
\usage{
sched_parse_teams(.data)
}
\arguments{
\item{.data}{The teams data read by \code{sched_load_teams()}}
}
\value{
A \code{tibble} containing one column for each team (named with that
team's letter designation) and one row per team member. Team member names
are standardized with
\code{\link[covidassign:sched_std_names]{sched_std_names()}}.
}
\description{
\code{sched_parse_teams()} parses the data returned from \code{sched_load_teams()} into
a standard format, with one column per team and one row per individual.
}
