% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sched_parse_teams.R
\name{sched_remove_phn}
\alias{sched_remove_phn}
\title{Remove "PHN" Placeholder From Nurse Role}
\usage{
sched_remove_phn(string)
}
\arguments{
\item{string}{A character vector}
}
\value{
A character vector with "PHN" removed when surrounded by only
symbols and/or numbers. Not case sensitive.
}
\description{
Remove "PHN" Placeholder From Nurse Role
}
