% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_addresses.R
\name{squish_addr}
\alias{squish_addr}
\title{"Squish" Address Variable Between Parsing Steps in the postmastr workflow}
\usage{
squish_addr(.data, .cols = "pm.address")
}
\arguments{
\item{.data}{An intermediate \code{tibble} in the postmastr workflow}

\item{.cols}{\verb{<tidy-select>} \code{character} columns to squish. If any of
these are not of type \code{character}, they are ignored with a warning.}
}
\value{
\code{.data} with squished character columns
}
\description{
\code{squish_addr()} removes extraneous whitespace from the columns specified by
\code{.cols}; this is typically the \code{pm.address} variable used in the postmastr
workflow.
}
\keyword{internal}
