% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/std_names.R
\name{str_invert_na}
\alias{str_invert_na}
\title{Invert \code{NA} String Replacement}
\usage{
str_invert_na(string)
}
\arguments{
\item{string}{A character vector}
}
\value{
The input \code{string} with \code{NA} replacements converted back to
\code{NA_character_}
}
\description{
\code{str_inv_replace_na()} performs the inverse operation of
\code{\link[stringr:str_replace_na]{str_replace_na()}}; that is, it converts
replaced \code{NA} values back to \code{NA_character_}. This only works if \code{NA} is
replaced with \code{"NA"} or a case-wise variant thereof (i.e. \code{"Na"}, \code{"na"}, or
\code{"nA"}). It is designed to invert the addition of "\verb{_[0-9]}" suffixes as
well.
}
\keyword{internal}
