% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/std_names.R
\name{str_remove_ordinal}
\alias{str_remove_ordinal}
\title{Remove Mixed Ordinal Numbers with Text Representation}
\usage{
str_remove_ordinal(string)
}
\arguments{
\item{string}{A character vector}

\item{num_paste}{A logical. If \code{TRUE}, the elements of larger numbers are
separated with spaces. If \code{FALSE}, the elements will be joined without
spaces.}

\item{A}{logical. If \code{TRUE}, ordinal numbers are removed from \code{string}}

\item{...}{Ignored.}
}
\description{
Removes mixed text/numeric ordinal numbers (i.e. "1st", "2nd", "23rd"). This
function was inspired by the \code{replace_ordinal()} function in the
textclean package, and re-uses pieces of its code.
}
\details{
Currently only implemented for ordinal values 1 through 100.
}
\keyword{internal}
