% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-team_schedules.R
\docType{data}
\name{team_schedules}
\alias{team_schedules}
\title{Team Schedules Used for Case Assignment}
\format{
A data frame with 7 rows and 3 variables:
\describe{
\item{team}{a character vector of investigation team letter designations}
\item{schedule}{a character vector of schedule types}
\item{anchor}{a \code{Date} vecotr of anchor dates for rotating schedules}
}
}
\source{
"Investigation Shift Schedules.xlsx"
}
\usage{
team_schedules
}
\description{
A dataset containing the case investigation teams, the schedule followed by
each team, and the anchor date for the rotating schedules.
}
\keyword{datasets}
