% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_json_redcap.R
\name{to_json_redcap}
\alias{to_json_redcap}
\title{Convert a Data Frame to JSON for Upload to REDcap}
\usage{
to_json_redcap(.data, size_limit = as.integer(0.9 * 3.2e+07), ...)
}
\arguments{
\item{.data}{A data frame or data frame extension}

\item{size_limit}{The maximum allowed size of the resulting JSON object}

\item{...}{Additional arguments to pass to
\code{\link[jsonlite:toJSON]{toJSON()}}}
}
\value{
A \code{json} character object
}
\description{
\code{to_json_redcap()} uses \code{\link[jsonlite:toJSON]{toJSON()}} to convert
a data frame to a JSON object for import into REDcap. It checks that the
size of the resultin object does not exceed \code{size_limit}, which (by default)
is 90\% of the maximum allowed upload size in the Shelby County REDcap.
}
