% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_assignments.R
\name{upload_assignments}
\alias{upload_assignments}
\title{Upload Assigned Cases to REDcap Case Assignment Project}
\usage{
upload_assignments(
  .data,
  api_token = Sys.getenv("redcap_NCA_token"),
  size_limit = as.integer(0.9 * 3.2e+07)
)
}
\arguments{
\item{.data}{Assigned cases data}

\item{api_token}{API token for REDcap project}
}
\value{
An httr \code{Response} w/ content indicating the number of uploaded
records
}
\description{
\code{upload_assignments()} uploads assigned cases to the Case Assignment REDcap
project, with \code{Investigator} and \verb{Date of Assignment} fields filled.
}
