% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-vctrs.R
\name{utils-vctrs}
\alias{utils-vctrs}
\title{vctrs: Vector Helpers}
\description{
The vctrs package defines new notions of prototype and size that are used to
provide tools for consistent and well-founded type-coercion and
size-recycling, and are in turn connected to ideas of type- and
size-stability useful for analyzing function interfaces. See the vctrs
site at \url{https://vctrs.r-lib.org/} for more information.
}
\details{
This package is imported directly to encourage development using the
standardized helpers it provides. This decision may be re-thought at a later
date.
}
