% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_assignments.R
\name{validate_assignments}
\alias{validate_assignments}
\title{Validate Case Assignments}
\usage{
validate_assignments(
  .data = assign_acns(),
  archive = TRUE,
  notify_to = c("Jesse.Smith@shelbycountytn.gov", "Karim.Gilani@shelbycountytn.gov"),
  cnd = c("warning", "error"),
  date = attr(.data, "date")
)
}
\arguments{
\item{.data}{Assignment data from \code{assign_acns()}}

\item{archive}{Should excluded investigators be archived?}

\item{notify_to}{Email addresses to notify of excluded investigators. Set to
\code{NULL} to disable notifications.}

\item{cnd}{Should excluded investigators trigger a warning or an error?}

\item{date}{The date assigned to the output \code{date_tbl}}
}
\value{
Excluded investigators in a \code{date_tbl}
}
\description{
\code{validate_assignments()} validates that cases were assigned to all scheduled
investigators. By default, it archives any excluded investigators and
notifies relevant parties of the omission.
}
