% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_upload.R
\name{validate_upload}
\alias{validate_upload}
\title{Validate Case Assignment Upload}
\usage{
validate_upload(
  .data,
  archive = TRUE,
  notify_to = c("Jesse.Smith@shelbycountytn.gov", "Karim.Gilani@shelbycountytn.gov"),
  cnd = c("warning", "error"),
  date = attr(.data, "date")
)
}
\arguments{
\item{.data}{Upload data from \code{upload_assigments()}}

\item{archive}{Should accepted and rejected cases be archived?}

\item{notify_to}{Email addresses to notify of case assignment. Set to \code{NULL}
to disable notification.}

\item{cnd}{Should rejected cases trigger a warning or an error?}

\item{date}{The date to use in notification}
}
\value{
The input data
}
\description{
\code{validate_upload()} validates that all assigned cases were uploaded. By
default, it archives accepted and rejected cases and notifies relevant
parties of the status. Note that, unlike \code{validate_assignments()}, a
notification is sent even if all cases are uploaded correctly.
}
