% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_redcap_cases.R
\name{download_redcap_cases}
\alias{download_redcap_cases}
\title{Download Cases from Case Assignment REDcap Project}
\usage{
download_redcap_cases(
  api_token = Sys.getenv("redcap_CA_token"),
  unassigned_only = TRUE
)
}
\arguments{
\item{api_token}{The API token for the Case Assignment REDcap project. The
default pulls this from the environment variable \code{redcap_CA_token}}

\item{unassigned_only}{Should only unassigned cases be downloaded?}
}
\value{
A tidy \code{tibble} with all variables in \code{character} format
}
\description{
\code{download_redcap_cases()} downloads records from the
\strong{Case Assignment} REDcap project. If \code{incl_assigned = FALSE}, it only
downloads unassigned cases; if \code{incl_assigned = TRUE}, it downloads all
cases.
}
