% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_acns.R
\name{translate_acns}
\alias{translate_acns}
\title{Translate ACNS Data to REDcap-Useable Format}
\usage{
translate_acns(
  .data = covidsms::prep_acns(assign = TRUE),
  date = lubridate::today(),
  days = 6L
)
}
\arguments{
\item{.data}{ACNS data. Runs \code{covidsms::prep_acns(assign = TRUE)} by default.}

\item{date}{The date to use for assignment}

\item{days}{Number of days back to consider valid. \code{test_date} older than
this will not be assigned.}
}
\value{
The data mapped to REDcap field names
}
\description{
\code{translate_acns()} maps fields from ACNS data to REDcap data and excludes
duplicates + cases older than \code{days} old.
}
